chore: new set init
