--- class/app.js
@@ -1,5 +1,0 @@
-export default class App {
-  constructor() {
-    console.log('hi!');
-  }
-}
