+++ function/event/triggerClickEvent.js
