+++ function/event/triggerDoubleClickEvent.js
