+++ function/event/triggerKeyupEvent.js
