+++ function/todoList/countTodoItem.js
