+++ function/todoList/removeTodoItem.js
